nlp-service/requirements-clean.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3

PyMuPDF==1.24.10
rapidocr-onnxruntime==1.4.4
onnxruntime==1.18.1
opencv-python==4.11.0.86
pillow==11.2.1

groq==0.11.0

huggingface-hub==0.30.2
sentence-transformers==3.0.1
torch==2.7.0

rank-bm25==0.2.2
numpy==1.26.4
scikit-learn==1.5.2

reportlab==4.4.10
